Fix bug where bought items end up in enderchest
Enderchests add a new player inventory list. This ended up being the
first list, so when you bought something it would end up in the
enderchest.

We fix this bug by adding only to the "main" list when buying items. It
didn't make sense to add to the crafting list or any other list anyway.