Disable parts of the economy if < 2 players online
When less than 2 players are online there are not enough people to react
to changes in the economy (e.g. increases in basic income, changing
prices). This can often lead to hyperinflation.

To fix this issue, I've disabled certain parts of the economy when less
than 2 players online. Namely:

1. No buying items.
2. No tax (since if there is one player, the is no incentive to set
resonable prices)
3. No income (prevents inflation)
4. No auctioning or bankruptcy
5. No "supply integration" (as the calculated values would be wrong if
no one is connected).